+++ packages/mask/src/plugins/Snapshot/SNSAdaptor/hooks/useVotes.ts
@@ -1,25 +1,13 @@
-import type { Proposal, ProposalIdentifier, VoteItem } from '../../types.js'
+import type { ProposalIdentifier, VoteItem } from '../../types.js'
 import { cache, use } from 'react'
 import { useProposal } from './useProposal.js'
 import { sumBy } from 'lodash-es'
 import { isSameAddress } from '@masknet/web3-shared-base'
+import { getScores } from '../../utils.js'
 
 const Request = cache(Suspender)
 export function useVotes(identifier: ProposalIdentifier, account?: string) {
     return use(Request(identifier.id, identifier.space, account))
-}
-
-function getScores(proposal: Proposal) {
-    const scores = []
-    for (let i = 0; i < proposal.choices.length; i += 1) {
-        const score: Record<string, number> = {}
-        for (const vote of proposal.votes) {
-            if (vote.vp_by_strategy[i] > 0) score[vote.voter.toLowerCase()] = vote.vp_by_strategy[i]
-        }
-        scores.push(score)
-    }
-
-    return scores
 }
 
 async function Suspender(id: ProposalIdentifier['id'], space: ProposalIdentifier['space'], account?: string) {
