+++ packages/mask/src/plugins/Snapshot/utils.ts
@@ -1,3 +1,5 @@
+import type { Proposal } from './types.js'
+
 export const getSnapshotVoteType = (type: string) => ({
     Vote: [
         {
@@ -26,3 +28,16 @@ export const getSnapshotVoteType = (type: string) => ({
         },
     ],
 })
+
+export function getScores(proposal: Proposal) {
+    const scores = []
+    for (let i = 0; i < proposal.choices.length; i += 1) {
+        const score: Record<string, number> = {}
+        for (const vote of proposal.votes) {
+            if (vote.vp_by_strategy[i] > 0) score[vote.voter.toLowerCase()] = vote.vp_by_strategy[i]
+        }
+        scores.push(score)
+    }
+
+    return scores
+}
